personal_assistant_web_app/backend/flask-server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bardapi==0.1.38
beautifulsoup4==4.12.2
blinker==1.7.0
browser-cookie3==0.19.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deep-translator==1.11.4
distro==1.8.0
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-texttospeech==2.14.2
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
lz4==4.3.2
MarkupSafe==2.1.3
openai==1.2.3
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.5.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
